Refactor sequential local placer modules
