--- src/transform/place_and_route/local_placer/sequential.rs
@@ -1,885 +1,0 @@
-use super::*;
-use crate::sequential::core::rs_latch_prefix_graph;
-use crate::world::World;
-
-pub(super) fn supports_sequential_primitive(sequential: &SequentialPrimitive) -> bool {
-    matches!(
-        sequential.sequential_type,
-        SequentialType::RsLatch | SequentialType::DLatch
-    ) && (sequential.rs_latch_core().is_some()
-        || !SequentialMacro::candidates(sequential).is_empty())
-}
-
-#[derive(Debug, Clone)]
-pub(super) struct RsLatchGatePlacement {
-    pub(super) world: World3D,
-    pub(super) q_torch: Position,
-    pub(super) q_cobble: Position,
-    pub(super) nq_torch: Position,
-    pub(super) nq_cobble: Position,
-}
-
-#[derive(Debug, Clone, Copy)]
-pub(super) struct RsLatchInputNodeIds {
-    pub(super) set: GraphNodeId,
-    pub(super) reset: GraphNodeId,
-}
-
-const SEQUENTIAL_SYNTHETIC_NODE_ID_STRIDE: usize = 4;
-
-// Prefix placement runs in its own graph, then publishes only the set/reset
-// endpoints back into the outer placement state for the RS latch core router.
-pub(super) fn rs_latch_input_node_ids(node_id: GraphNodeId) -> RsLatchInputNodeIds {
-    let base = usize::MAX - node_id * SEQUENTIAL_SYNTHETIC_NODE_ID_STRIDE;
-    RsLatchInputNodeIds {
-        set: base - 1,
-        reset: base - 2,
-    }
-}
-
-struct RsLatchPrefixPlan {
-    graph: LogicGraph,
-    input_mappings: Vec<(GraphNodeId, GraphNodeId)>,
-    input_node_ids_by_port: HashMap<String, GraphNodeId>,
-    input_positions_by_port: HashMap<String, Position>,
-    set_source: GraphNodeId,
-    reset_source: GraphNodeId,
-    rs_input_nodes: RsLatchInputNodeIds,
-}
-
-impl RsLatchPrefixPlan {
-    fn build(
-        node: &GraphNode,
-        sequential: &SequentialPrimitive,
-        state: &PlacementState,
-    ) -> Option<Self> {
-        let core = sequential.rs_latch_core()?;
-        let graph = rs_latch_prefix_graph(&sequential.inner_graph, &core)
-            .and_then(|graph| graph.prepare_place().ok())?;
-        let set_source = output_source_node_id(&graph, "s")?;
-        let reset_source = output_source_node_id(&graph, "r")?;
-
-        let mut input_mappings = Vec::new();
-        let mut input_node_ids_by_port = HashMap::new();
-        let mut input_positions_by_port = HashMap::new();
-        for prefix_input in graph.nodes.iter().filter_map(|node| match &node.kind {
-            GraphNodeKind::Input(name) => Some((node.id, name.as_str())),
-            _ => None,
-        }) {
-            let outer_input = outer_input_node_id(node, sequential, prefix_input.1)?;
-            let position = state.node_position(outer_input)?;
-            input_mappings.push((prefix_input.0, outer_input));
-            input_node_ids_by_port.insert(prefix_input.1.to_owned(), outer_input);
-            input_positions_by_port.insert(prefix_input.1.to_owned(), position);
-        }
-
-        Some(Self {
-            graph,
-            input_mappings,
-            input_node_ids_by_port,
-            input_positions_by_port,
-            set_source,
-            reset_source,
-            rs_input_nodes: rs_latch_input_node_ids(node.id),
-        })
-    }
-
-    fn rs_node_inputs(&self) -> Vec<GraphNodeId> {
-        vec![self.rs_input_nodes.set, self.rs_input_nodes.reset]
-    }
-
-    fn input_node_id(&self, port: &str) -> Option<GraphNodeId> {
-        self.input_node_ids_by_port.get(port).copied()
-    }
-
-    fn input_position(&self, port: &str) -> Option<Position> {
-        self.input_positions_by_port.get(port).copied()
-    }
-
-    fn d_latch_input_node_ids(&self) -> Option<(GraphNodeId, GraphNodeId)> {
-        Some((self.input_node_id("d")?, self.input_node_id("en")?))
-    }
-
-    fn source_positions(&self, state: &PlacementState) -> Option<(Position, Position)> {
-        Some((
-            state.node_position(self.set_source)?,
-            state.node_position(self.reset_source)?,
-        ))
-    }
-
-    fn publish_rs_inputs(
-        &self,
-        state: &mut PlacementState,
-        set_position: Position,
-        reset_position: Position,
-    ) {
-        state.set_node_position(self.rs_input_nodes.set, set_position);
-        state.set_node_position(self.rs_input_nodes.reset, reset_position);
-    }
-}
-
-pub(super) fn generate_d_latch_gate_routes(
-    config: &LocalPlacerConfig,
-    node: &GraphNode,
-    sequential: &SequentialPrimitive,
-    world: &World3D,
-    state: &PlacementState,
-) -> PlacerQueue {
-    let Some(prefix_plan) = RsLatchPrefixPlan::build(node, sequential, state) else {
-        return Vec::new();
-    };
-    let Some((data_input, enable_input)) = prefix_plan.d_latch_input_node_ids() else {
-        return Vec::new();
-    };
-    let rs_sequential = SequentialPrimitive::new(
-        SequentialType::RsLatch,
-        vec!["s".to_owned(), "r".to_owned()],
-        sequential.output_ports.clone(),
-    );
-    let rs_node = GraphNode {
-        id: node.id,
-        kind: GraphNodeKind::Sequential(rs_sequential.clone()),
-        inputs: prefix_plan.rs_node_inputs(),
-        ..Default::default()
-    };
-
-    let queue = place_rs_latch_prefix(config, world, &prefix_plan, state);
-    let queue = retain_valid_d_latch_prefix(config, queue, &prefix_plan);
-
-    let mut routed = Vec::new();
-    let limit = take_sampling_limit(config.step_sampling_policy);
-    let rs_config = LocalPlacerConfig {
-        step_sampling_policy: SamplingPolicy::Random(32),
-        route_step_sampling_policy: SamplingPolicy::Random(32),
-        not_route_step_sampling_policy: SamplingPolicy::Random(32),
-        max_not_route_step: 8,
-        max_route_step: 8,
-        ..*config
-    };
-    for (world, prefix_state) in queue {
-        let Some((set_position, reset_position)) = prefix_plan.source_positions(&prefix_state)
-        else {
-            continue;
-        };
-        let mut state = state.clone();
-        prefix_plan.publish_rs_inputs(&mut state, set_position, reset_position);
-        for (world, state) in
-            generate_rs_latch_gate_routes(&rs_config, &rs_node, &rs_sequential, &world, &state)
-        {
-            let Some(q) = state.port_position(node.id, "q") else {
-                continue;
-            };
-            let Some(nq) = state.port_position(node.id, "nq") else {
-                continue;
-            };
-            if std::panic::catch_unwind(|| {
-                d_latch_candidate_behaves(&world, data_input, enable_input, &state, q, nq)
-            })
-            .unwrap_or(false)
-            {
-                routed.push((world, state));
-                break;
-            }
-        }
-        if limit.is_some_and(|limit| routed.len() >= limit) {
-            break;
-        }
-        if !routed.is_empty() {
-            break;
-        }
-    }
-    sample_d_latch_step(config, 6, routed)
-}
-
-fn place_rs_latch_prefix(
-    config: &LocalPlacerConfig,
-    world: &World3D,
-    prefix_plan: &RsLatchPrefixPlan,
-    outer_state: &PlacementState,
-) -> PlacerQueue {
-    let mut prefix_state = PlacementState::default();
-    for &(prefix_input, outer_input) in &prefix_plan.input_mappings {
-        let Some(position) = outer_state.node_position(outer_input) else {
-            return Vec::new();
-        };
-        prefix_state.set_node_position(prefix_input, position);
-    }
-
-    let Ok(prefix_placer) = LocalPlacer::new(prefix_plan.graph.clone(), *config) else {
-        return Vec::new();
-    };
-    prefix_placer.generate_queue_from(vec![(world.clone(), prefix_state)], None, None)
-}
-
-fn retain_valid_d_latch_prefix(
-    config: &LocalPlacerConfig,
-    queue: PlacerQueue,
-    prefix_plan: &RsLatchPrefixPlan,
-) -> PlacerQueue {
-    let Some(data) = prefix_plan.input_position("d") else {
-        return Vec::new();
-    };
-    let Some(enable) = prefix_plan.input_position("en") else {
-        return Vec::new();
-    };
-    let queue = queue
-        .into_iter()
-        .filter(|(world, state)| {
-            d_latch_prefix_behaves(
-                world,
-                data,
-                enable,
-                prefix_plan.set_source,
-                prefix_plan.reset_source,
-                state,
-            )
-        })
-        .collect_vec();
-    SamplingPolicy::Random(64).sample_with_seed(queue, config.sampling_seed(7, 6))
-}
-
-fn d_latch_prefix_behaves(
-    world: &World3D,
-    data: Position,
-    enable: Position,
-    set_node_id: GraphNodeId,
-    reset_node_id: GraphNodeId,
-    state: &PlacementState,
-) -> bool {
-    let Some(set) = state.node_position(set_node_id) else {
-        return false;
-    };
-    let Some(reset) = state.node_position(reset_node_id) else {
-        return false;
-    };
-
-    [(false, false), (false, true), (true, false), (true, true)]
-        .into_iter()
-        .all(|(data_value, enable_value)| {
-            d_latch_input_gate_values(world, data, enable, set, reset, data_value, enable_value)
-                .is_some_and(|(set_value, reset_value)| {
-                    set_value == (data_value && enable_value)
-                        && reset_value == (!data_value && enable_value)
-                })
-        })
-}
-
-fn d_latch_input_gate_values(
-    world: &World3D,
-    data: Position,
-    enable: Position,
-    set: Position,
-    reset: Position,
-    data_value: bool,
-    enable_value: bool,
-) -> Option<(bool, bool)> {
-    let mut world = pad_world_top_for_simulation(world, 2);
-    world[data].kind = BlockKind::Switch { is_on: data_value };
-    world[enable].kind = BlockKind::Switch {
-        is_on: enable_value,
-    };
-    world.initialize_redstone_states();
-    let world = World::from(&world);
-    let sim = Simulator::from_with_limits_and_trace(&world, 64, 20_000, 0).ok()?;
-    let set_value = match sim.world()[set].kind {
-        BlockKind::Torch { is_on } => is_on,
-        _ => return None,
-    };
-    let reset_value = match sim.world()[reset].kind {
-        BlockKind::Torch { is_on } => is_on,
-        _ => return None,
-    };
-    Some((set_value, reset_value))
-}
-
-fn d_latch_candidate_behaves(
-    world: &World3D,
-    data_node_id: GraphNodeId,
-    enable_node_id: GraphNodeId,
-    state: &PlacementState,
-    q: Position,
-    nq: Position,
-) -> bool {
-    let Some(data) = state.node_position(data_node_id) else {
-        return false;
-    };
-    let Some(enable) = state.node_position(enable_node_id) else {
-        return false;
-    };
-
-    let mut reset_world = pad_world_top_for_simulation(world, 2);
-    reset_world[enable].kind = BlockKind::Switch { is_on: true };
-    reset_world[data].kind = BlockKind::Switch { is_on: false };
-    reset_world.initialize_redstone_states();
-    let world = World::from(&reset_world);
-    let Ok(mut sim) = Simulator::from_with_limits_and_trace(&world, 64, 20_000, 0) else {
-        return false;
-    };
-
-    if torch_is_on_in_world3d(sim.world(), q) || !torch_is_on_in_world3d(sim.world(), nq) {
-        return false;
-    }
-    if sim
-        .change_state_with_limits(vec![(enable, false)], 64, 20_000)
-        .is_err()
-    {
-        return false;
-    }
-    if sim
-        .change_state_with_limits(vec![(data, true)], 64, 20_000)
-        .is_err()
-    {
-        return false;
-    }
-    if torch_is_on_in_world3d(sim.world(), q) || !torch_is_on_in_world3d(sim.world(), nq) {
-        return false;
-    }
-    if sim
-        .change_state_with_limits(vec![(enable, true)], 64, 20_000)
-        .is_err()
-    {
-        return false;
-    }
-    if !torch_is_on_in_world3d(sim.world(), q) || torch_is_on_in_world3d(sim.world(), nq) {
-        return false;
-    }
-    if sim
-        .change_state_with_limits(vec![(enable, false)], 64, 20_000)
-        .is_err()
-    {
-        return false;
-    }
-    if sim
-        .change_state_with_limits(vec![(data, false)], 64, 20_000)
-        .is_err()
-    {
-        return false;
-    }
-    if !torch_is_on_in_world3d(sim.world(), q) || torch_is_on_in_world3d(sim.world(), nq) {
-        return false;
-    }
-    if sim
-        .change_state_with_limits(vec![(enable, true)], 64, 20_000)
-        .is_err()
-    {
-        return false;
-    }
-
-    !torch_is_on_in_world3d(sim.world(), q) && torch_is_on_in_world3d(sim.world(), nq)
-}
-
-fn torch_is_on_in_world3d(world: &World3D, position: Position) -> bool {
-    matches!(world[position].kind, BlockKind::Torch { is_on: true })
-}
-
-fn pad_world_top_for_simulation(world: &World3D, extra_z: usize) -> World3D {
-    let mut padded = World3D::new(DimSize(world.size.0, world.size.1, world.size.2 + extra_z));
-    for (position, mut block) in world.iter_block() {
-        if block.kind.is_cobble() {
-            block.kind = BlockKind::Cobble {
-                on_count: 0,
-                on_base_count: 0,
-            };
-        }
-        padded[position] = block;
-    }
-    padded
-}
-
-fn outer_input_node_id(
-    node: &GraphNode,
-    sequential: &SequentialPrimitive,
-    port: &str,
-) -> Option<GraphNodeId> {
-    node.inputs
-        .iter()
-        .zip(&sequential.input_ports)
-        .find_map(|(input_node, input_port)| (input_port == port).then_some(*input_node))
-}
-
-fn output_source_node_id(graph: &LogicGraph, output_name: &str) -> Option<GraphNodeId> {
-    graph.nodes.iter().find_map(|node| {
-        (matches!(&node.kind, GraphNodeKind::Output(name) if name == output_name)
-            && node.inputs.len() == 1)
-            .then(|| node.inputs[0])
-    })
-}
-
-fn sample_d_latch_step(
-    config: &LocalPlacerConfig,
-    step: usize,
-    candidates: PlacerQueue,
-) -> PlacerQueue {
-    config
-        .step_sampling_policy
-        .sample_with_seed(candidates, config.sampling_seed(7, step))
-}
-
-fn take_sampling_limit(policy: SamplingPolicy) -> Option<usize> {
-    match policy {
-        SamplingPolicy::Take(count) => Some(count),
-        SamplingPolicy::None | SamplingPolicy::Random(_) => None,
-    }
-}
-
-pub(super) fn generate_rs_latch_gate_routes(
-    config: &LocalPlacerConfig,
-    node: &GraphNode,
-    sequential: &SequentialPrimitive,
-    world: &World3D,
-    state: &PlacementState,
-) -> PlacerQueue {
-    let raw_pairs = generate_rs_latch_not_pairs(world);
-    let pairs = select_rs_latch_not_pairs(config, node, sequential, state, raw_pairs.clone());
-    let mut routed = Vec::new();
-    let limit = take_sampling_limit(config.step_sampling_policy);
-    for pair in pairs {
-        for placed in route_rs_latch_branches(config, node, sequential, state, pair) {
-            let mut state = state.clone();
-            state.set_port_position(node.id, "q".to_owned(), placed.q_torch);
-            state.set_port_position(node.id, "nq".to_owned(), placed.nq_torch);
-            let primary_output = sequential
-                .output_ports
-                .first()
-                .map(String::as_str)
-                .unwrap_or("q");
-            let primary_position = match primary_output {
-                "q" => placed.q_torch,
-                "nq" => placed.nq_torch,
-                _ => placed.q_torch,
-            };
-            state.set_node_position(node.id, primary_position);
-            routed.push((placed.world, state));
-            if limit.is_some_and(|limit| routed.len() >= limit) {
-                break;
-            }
-        }
-        if limit.is_some_and(|limit| routed.len() >= limit) {
-            break;
-        }
-    }
-
-    if routed.is_empty() {
-        generate_sequential_macro_routes(config, node, sequential, world, state)
-    } else {
-        routed
-    }
-}
-
-pub(super) fn select_rs_latch_not_pairs(
-    config: &LocalPlacerConfig,
-    node: &GraphNode,
-    sequential: &SequentialPrimitive,
-    state: &PlacementState,
-    mut pairs: Vec<RsLatchGatePlacement>,
-) -> Vec<RsLatchGatePlacement> {
-    let Some(set_source) = rs_latch_input_source(node, sequential, state, "s") else {
-        return config
-            .step_sampling_policy
-            .sample_with_seed(pairs, config.sampling_seed(4, 0));
-    };
-    let Some(reset_source) = rs_latch_input_source(node, sequential, state, "r") else {
-        return Vec::new();
-    };
-
-    pairs.sort_by_key(|placed| {
-        let reset_distance = reset_source.manhattan_distance(&placed.q_cobble);
-        let set_distance = set_source.manhattan_distance(&placed.nq_cobble);
-        let input_block_penalty = if lies_between(reset_source, placed.q_cobble, placed.q_torch) {
-            1_000
-        } else {
-            0
-        } + if lies_between(set_source, placed.nq_cobble, placed.nq_torch)
-        {
-            1_000
-        } else {
-            0
-        };
-        let input_space_penalty = (reset_distance.abs_diff(set_distance) * 10)
-            + usize::from(reset_distance < 2) * 100
-            + usize::from(set_distance < 2) * 100;
-        reset_distance
-            + set_distance
-            + placed.q_torch.manhattan_distance(&placed.nq_torch)
-            + input_block_penalty
-            + input_space_penalty
-    });
-
-    match config.step_sampling_policy {
-        SamplingPolicy::None => pairs,
-        SamplingPolicy::Take(count) | SamplingPolicy::Random(count) => {
-            pairs.into_iter().take(count).collect()
-        }
-    }
-}
-
-fn lies_between(start: Position, end: Position, position: Position) -> bool {
-    start.manhattan_distance(&position) + position.manhattan_distance(&end)
-        == start.manhattan_distance(&end)
-        && position != start
-        && position != end
-}
-
-pub(super) fn route_rs_latch_branches(
-    config: &LocalPlacerConfig,
-    node: &GraphNode,
-    sequential: &SequentialPrimitive,
-    state: &PlacementState,
-    placed: RsLatchGatePlacement,
-) -> Vec<RsLatchGatePlacement> {
-    let Some(set_source) = rs_latch_input_source(node, sequential, state, "s") else {
-        return route_rs_latch_feedback(config, placed);
-    };
-    let Some(reset_source) = rs_latch_input_source(node, sequential, state, "r") else {
-        return Vec::new();
-    };
-
-    let q_torch = placed.q_torch;
-    let q_cobble = placed.q_cobble;
-    let nq_torch = placed.nq_torch;
-    let nq_cobble = placed.nq_cobble;
-    route_rs_latch_signals_to_cobble(config, placed, nq_torch, reset_source, q_torch, q_cobble)
-        .into_iter()
-        .flat_map(|placed| {
-            route_rs_latch_signals_to_cobble(
-                config, placed, q_torch, set_source, nq_torch, nq_cobble,
-            )
-        })
-        .collect()
-}
-
-fn route_rs_latch_signals_to_cobble(
-    config: &LocalPlacerConfig,
-    placed: RsLatchGatePlacement,
-    first_source: Position,
-    second_source: Position,
-    target_torch: Position,
-    target_cobble: Position,
-) -> Vec<RsLatchGatePlacement> {
-    let worlds = generate_two_routes_to_cobble(
-        config,
-        &placed.world,
-        first_source,
-        second_source,
-        target_torch,
-        target_cobble,
-    );
-    config
-        .route_step_sampling_policy
-        .sample_with_seed(
-            worlds,
-            config.sampling_seed(8, target_cobble.0 + target_cobble.1),
-        )
-        .into_iter()
-        .map(|world| RsLatchGatePlacement { world, ..placed })
-        .collect()
-}
-
-fn generate_two_routes_to_cobble(
-    config: &LocalPlacerConfig,
-    world: &World3D,
-    first_source: Position,
-    second_source: Position,
-    target_torch: Position,
-    target_cobble: Position,
-) -> Vec<World3D> {
-    let first_routes = generate_routes_to_cobble_with_paths(
-        config,
-        world,
-        first_source,
-        target_torch,
-        target_cobble,
-    );
-    let second_routes = generate_routes_to_cobble_with_paths(
-        config,
-        world,
-        second_source,
-        target_torch,
-        target_cobble,
-    );
-
-    let first_then_second = first_routes.iter().flat_map(|(first_world, first_path)| {
-        let network = redstone_route_network(first_world, first_path, target_cobble);
-        generate_routes_to_cobble_or_network(
-            config,
-            first_world,
-            second_source,
-            target_torch,
-            target_cobble,
-            &network,
-        )
-        .into_iter()
-        .map(|(world, _)| world)
-        .collect_vec()
-    });
-
-    let second_then_first = second_routes
-        .iter()
-        .flat_map(|(second_world, second_path)| {
-            let network = redstone_route_network(second_world, second_path, target_cobble);
-            generate_routes_to_cobble_or_network(
-                config,
-                second_world,
-                first_source,
-                target_torch,
-                target_cobble,
-                &network,
-            )
-            .into_iter()
-            .map(|(world, _)| world)
-            .collect_vec()
-        });
-
-    first_then_second.chain(second_then_first).collect()
-}
-
-fn redstone_route_network(
-    world: &World3D,
-    path: &[Position],
-    target_cobble: Position,
-) -> HashSet<Position> {
-    path.iter()
-        .copied()
-        .filter(|&position| {
-            world[position].kind.is_redstone()
-                && redstone_powers_cobble(world, position, target_cobble)
-        })
-        .collect()
-}
-
-fn rs_latch_input_source(
-    node: &GraphNode,
-    sequential: &SequentialPrimitive,
-    state: &PlacementState,
-    port: &str,
-) -> Option<Position> {
-    node.inputs
-        .iter()
-        .zip(&sequential.input_ports)
-        .find_map(|(input_node, input_port)| {
-            (input_port == port)
-                .then(|| state.node_position(*input_node))
-                .flatten()
-        })
-}
-
-pub(super) fn generate_rs_latch_not_pairs(world: &World3D) -> Vec<RsLatchGatePlacement> {
-    let cardinal = [
-        Direction::East,
-        Direction::West,
-        Direction::South,
-        Direction::North,
-    ];
-    let support_positions = iproduct!(0..world.size.0, 0..world.size.1, 1..world.size.2)
-        .map(|(x, y, z)| Position(x, y, z))
-        .collect_vec();
-    let mut candidates = Vec::new();
-
-    for &q_direction in &cardinal {
-        for q_cobble in &support_positions {
-            let Some(q_torch) = q_cobble.walk(q_direction.inverse()) else {
-                continue;
-            };
-            if !world.size.bound_on(q_torch) {
-                continue;
-            }
-            let q_torch_block = Block {
-                kind: BlockKind::Torch { is_on: false },
-                direction: q_direction,
-            };
-            let Some((q_world, q_torch, q_cobble)) =
-                place_torch_with_cobble(world, q_torch_block, q_torch)
-            else {
-                continue;
-            };
-
-            for &nq_direction in &cardinal {
-                for nq_cobble in &support_positions {
-                    if q_cobble == *nq_cobble || q_cobble.2 != nq_cobble.2 {
-                        continue;
-                    }
-                    let support_distance = q_cobble.manhattan_distance(nq_cobble);
-                    if !(2..=3).contains(&support_distance) || q_direction.inverse() != nq_direction
-                    {
-                        continue;
-                    }
-                    let Some(nq_torch) = nq_cobble.walk(nq_direction.inverse()) else {
-                        continue;
-                    };
-                    if !q_world.size.bound_on(nq_torch)
-                        || q_torch == nq_torch
-                        || !(3..=5).contains(&q_torch.manhattan_distance(&nq_torch))
-                    {
-                        continue;
-                    }
-                    let nq_torch_block = Block {
-                        kind: BlockKind::Torch { is_on: false },
-                        direction: nq_direction,
-                    };
-                    let Some((world, nq_torch, nq_cobble)) =
-                        place_torch_with_cobble(&q_world, nq_torch_block, nq_torch)
-                    else {
-                        continue;
-                    };
-                    candidates.push(RsLatchGatePlacement {
-                        world,
-                        q_torch,
-                        q_cobble,
-                        nq_torch,
-                        nq_cobble,
-                    });
-                }
-            }
-        }
-    }
-
-    candidates
-}
-
-fn route_rs_latch_feedback(
-    config: &LocalPlacerConfig,
-    placed: RsLatchGatePlacement,
-) -> Vec<RsLatchGatePlacement> {
-    generate_routes_to_cobble(
-        config,
-        &placed.world,
-        placed.nq_torch,
-        placed.q_torch,
-        placed.q_cobble,
-    )
-    .into_iter()
-    .flat_map(|(world, _)| {
-        generate_routes_to_cobble(
-            config,
-            &world,
-            placed.q_torch,
-            placed.nq_torch,
-            placed.nq_cobble,
-        )
-        .into_iter()
-        .map(|(world, _)| RsLatchGatePlacement {
-            world,
-            q_torch: placed.q_torch,
-            q_cobble: placed.q_cobble,
-            nq_torch: placed.nq_torch,
-            nq_cobble: placed.nq_cobble,
-        })
-        .collect_vec()
-    })
-    .collect()
-}
-
-pub(super) fn generate_sequential_macro_routes(
-    config: &LocalPlacerConfig,
-    node: &GraphNode,
-    sequential: &SequentialPrimitive,
-    world: &World3D,
-    state: &PlacementState,
-) -> PlacerQueue {
-    SequentialMacro::candidates(sequential)
-        .into_iter()
-        .flat_map(|candidate| {
-            iproduct!(0..world.size.0, 0..world.size.1, 0..world.size.2)
-                .map(|(x, y, z)| Position(x, y, z))
-                .filter_map(move |anchor| place_sequential_macro(world, &candidate, anchor))
-                .flat_map(|placed| {
-                    route_sequential_inputs(config, node, sequential, state, &placed)
-                        .into_iter()
-                        .map(|world| {
-                            let mut state = state.clone();
-                            for (output_port, position) in &placed.output_ports {
-                                state.set_port_position(node.id, output_port.to_owned(), *position);
-                            }
-                            let primary_output = sequential
-                                .output_ports
-                                .first()
-                                .unwrap_or(&placed.primary_output_port);
-                            let primary_position = placed.output_ports[primary_output];
-                            state.set_node_position(node.id, primary_position);
-                            (world, state)
-                        })
-                        .collect_vec()
-                })
-                .collect_vec()
-        })
-        .collect()
-}
-
-#[derive(Debug, Clone)]
-pub(super) struct PlacedSequentialMacro {
-    world: World3D,
-    input_ports: HashMap<String, Position>,
-    output_ports: HashMap<String, Position>,
-    primary_output_port: String,
-}
-
-pub(super) fn place_sequential_macro(
-    world: &World3D,
-    candidate: &SequentialMacro,
-    anchor: Position,
-) -> Option<PlacedSequentialMacro> {
-    let mut new_world = world.clone();
-    for (relative, block) in candidate.world.iter_block() {
-        let position = translate_position(anchor, relative)?;
-        if !world.size.bound_on(position) || !world[position].kind.is_air() {
-            return None;
-        }
-        new_world[position] = block;
-    }
-    new_world.initialize_redstone_states();
-
-    Some(PlacedSequentialMacro {
-        world: new_world,
-        input_ports: translate_ports(anchor, &candidate.input_ports)?,
-        output_ports: translate_ports(anchor, &candidate.output_ports)?,
-        primary_output_port: candidate.primary_output_port.clone(),
-    })
-}
-
-pub(super) fn route_sequential_inputs(
-    config: &LocalPlacerConfig,
-    node: &GraphNode,
-    sequential: &SequentialPrimitive,
-    state: &PlacementState,
-    placed: &PlacedSequentialMacro,
-) -> Vec<World3D> {
-    let mut worlds = vec![placed.world.clone()];
-    for (input_node, input_port) in node.inputs.iter().zip(&sequential.input_ports) {
-        let Some(source) = state.node_position(*input_node) else {
-            return Vec::new();
-        };
-        let Some(&target) = placed.input_ports.get(input_port) else {
-            return Vec::new();
-        };
-
-        worlds = worlds
-            .into_iter()
-            .flat_map(|world| {
-                generate_or_routes(config, &world, source, target)
-                    .routes
-                    .into_iter()
-                    .map(|(world, _)| world)
-                    .collect_vec()
-            })
-            .collect_vec();
-    }
-    worlds
-}
-
-fn translate_ports(
-    anchor: Position,
-    ports: &HashMap<String, Position>,
-) -> Option<HashMap<String, Position>> {
-    ports
-        .iter()
-        .map(|(port, position)| Some((port.clone(), translate_position(anchor, *position)?)))
-        .collect()
-}
-
-fn translate_position(anchor: Position, relative: Position) -> Option<Position> {
-    Some(Position(
-        anchor.0.checked_add(relative.0)?,
-        anchor.1.checked_add(relative.1)?,
-        anchor.2.checked_add(relative.2)?,
-    ))
-}
